Cast via uintptr_t when converting between integers and pointers
When casting from integer to pointer, promoting the integer to (u)intptr_t
will ensure that the resulting type can be converted to a pointer without
problems. These two casts changed in this commit trigger a warning when
building for CHERI-enabled architectures such as Arm Morello. This is not
just limited to CHERI, the cast from xcb_pixmap_t (uint32_t) to void*
should also be flagged by -Wint-to-void-pointer-cast when using Clang,
however, it appears that warning only handles C-style casts, and not
reinterpret_cast (https://github.com/llvm/llvm-project/issues/53964).

+++ src/backends/x11/standalone/eglbackend.cpp
@@ -228,8 +228,11 @@ bool EglPixmapTexturePrivate::create(SurfacePixmapX11 *pixmap)
         EGL_IMAGE_PRESERVED_KHR, EGL_TRUE,
         EGL_NONE
     };
-    m_image = eglCreateImageKHR(m_backend->eglDisplay(), EGL_NO_CONTEXT, EGL_NATIVE_PIXMAP_KHR,
-                                reinterpret_cast<EGLClientBuffer>(nativePixmap), attribs);
+    m_image = eglCreateImageKHR(m_backend->eglDisplay(),
+                                EGL_NO_CONTEXT,
+                                EGL_NATIVE_PIXMAP_KHR,
+                                reinterpret_cast<EGLClientBuffer>(static_cast<uintptr_t>(nativePixmap)),
+                                attribs);
 
     if (EGL_NO_IMAGE_KHR == m_image) {
         qCDebug(KWIN_CORE) << "failed to create egl image";
